reading_assessment/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pyttsx3==2.90
gTTS==2.3.2
pyaudio==0.2.14
numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
python-dotenv==1.0.0
requests==2.32.3
tensorflow==2.12.0
scipy==1.15.2 